Can now include members and groups with JSON
When sending the JSON response object,
the original list of members and groups
are ignored, and a specifically crafted
object (with only index numbers
and the group/member names) is
sent in its place.

This way we can decide what to send in response
while avoiding the infinite loop problem.