JSON response only contains names
Simplified the JSON representation of members
and groups, so that it now only prints the
name of the users, and the name of the groups.

toString methods now also handle the case
where they are newly created and dont
have any values, so the printing of
members and groups are formatted properly.