User passwords are now properly hashed
Added dependencies for hashing password using Argon2.
Added method for hashing a given password.
Added a test for verifying that a password is hashed.
Added a test to validate that changing the password
result in a hash that can be decrypted to the new password
(check that password was changed).